requirements.txt
Pillow
customtkinter
requests
pandas
